% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{addMajorTrans}
\alias{addMajorTrans}
\title{Add Major Trans Info to Existing abundance and count dataframes}
\usage{
addMajorTrans(
  genestouse,
  abGeneTempF,
  cntGeneTempF,
  abDatasets,
  CompMI = FALSE
)
}
\arguments{
\item{genestouse}{A list of genenames that are in use for the compositional analysis.}

\item{abGeneTempF}{data.frame of abundance information that will have the major transcript info added to it}

\item{cntGeneTempF}{data.frame of count information that will have the major transcript info added to it. Usually also contains length information.}

\item{abDatasets}{list of length genestouse containing genewise information per sample.  Also contains the major transcript information as an attribute.  Generally comes from the output of \code{\link{generateData}}.}

\item{CompMI}{is a TRUE/FALSE corresponding to whether datasets for the multiple imputation based analysis are being used.  This will add columns for transcripts that may be missing in the inferential replicates that were't missing in the non-inferential replicate data.  Default is FALSE.}
}
\value{
a list containing abGene and cntGene dataframes with the major transcript information added
}
\description{
\code{addMajorTrans} adds a binary indicator of the major trans for a given gene
}
\details{
\code{addMajorTrans} adds a binary indicator variable that takes a value of 1 if
that transcript is the major transcript for its gene (and 0 otherwise).  The major transcript
is the transcript that has the highest average expression proportion across all samples
for the given gene.  The major transcripts are computed in \code{\link{generateData}} and input
to this function via the abDatasets file.
}
