% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cntsToTPM}
\alias{cntsToTPM}
\title{Compute TPMs from counts and effective lengths}
\usage{
cntsToTPM(cnts, nsamp, len = NULL, samps = NULL)
}
\arguments{
\item{cnts}{is a dataframe of counts that has thecolum nnames Sample1Cnt, Sample2Cnt, ... and a column for tx_id.  Needs to also contain the (effective) length information Sample1Len, Sample2Len, ... unless the argument len is also specified}

\item{nsamp}{is the number of samples/biological replicates}

\item{len}{is an optional dataframe of length information.  len has (nsamp+1) columns with columnnames Sample1Len, Sample2Len, ... anda column for tx_id. Not needed if length information is contained in cnts.}

\item{samps}{is an optional vector containing the sample names.  Need to specify this if sample names are not just paste0("Sample", 1:nsamp) without any missing.}
}
\value{
data.frame of TPM information, with column names Sample1TPM, Sample2TPM, etc.
}
\description{
\code{cntsToTPM} computes TPMs from counts and lengths.  Rownames need
}
