% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CorrectLowExpression}
\alias{CorrectLowExpression}
\title{Correct sample/gene combinations that have expression values of 0 or close to 0 to stabilize results}
\usage{
CorrectLowExpression(y, CLEParam = 0.05)
}
\arguments{
\item{y}{is the data for the current gene/sample combination}

\item{CLEParam}{is the parameter (betwen 0 and 1) that controls the correction threshold (see details of \code{\link{CorrectLowExpression}} for more information)}
}
\description{
Correct sample/gene combinations that have expression values of 0 or close to 0 to stabilize results
}
\details{
The CLEParam parameter a works as follows: any TPM value that is less than `100*a' percent of the total gene-level
expression for the sample is replaced by `100*a' percent of this expression.  Mathematically, let \eqn{T_{ij}} be the TPM value for
transcript $j=1,..., D$ for sample $i = 1,..., n$ within a given gene with $D$ transcripts.  Any \deqn{T_{ij} < a * (T_{i1}+...+T_{iD})} will be replaced by \deqn{a * (T_{i1}+...+T_{iD})}.
This procedure results in relative transcript abundance fractions (RTAFs)
being zero only when every \eqn{T_{ij}} is equal to zero.  The value \eqn{a} could be increased or decreased to result in more or less modification to
the observed TPM values.  As $a$ increases, the proportions are driven closer to each other, with each converging towards \eqn{(1/D)} as \eqn{a} converges to 1
(and each equal to \eqn{(1/D)} for \eqn{a > 1}).  We find \deqn{a=0.05} is a good compromise that is large enough to stabilize the ilr coordinates sufficiently while
additionally not over-modifying the observed data, and use this value by default for all \emph{CompDTU} and \emph{CompDTUme} results. We recommend keeping this value at the default of 0.05.
}
