% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loadRData}
\alias{loadRData}
\title{Import selected object from a .RData file}
\usage{
loadRData(filePath, objNameToGet = NULL)
}
\arguments{
\item{filePath}{The path of the .RData object to load}

\item{objNameToGet}{The name of the object to extract from the specified .RData file.
If left to the default value of NULL the function will return the first object alphabetically so specify this argument if
more than one object is saved within the .RData object.}
}
\value{
An R object
}
\description{
\code{loadRData} imports selected object from a .RData file and returns only that object.  Modified from code from stackexchange user ricardo at this link https://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file
}
\details{
Modified from code from stackexchange user ricardo at this link https://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file
}
