% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SaveFullinfRepDat}
\alias{SaveFullinfRepDat}
\title{Save the datasets that contain data across all inferential replicates and samples for a particular subset of genes}
\usage{
SaveFullinfRepDat(
  SalmonFilesDir,
  save_dir,
  GibbsSamps,
  filteredgenenames,
  cntGene,
  key,
  curr_part_num,
  DRIMSeqFiltering = TRUE
)
}
\arguments{
\item{SalmonFilesDir}{is the directory the Salmon quantification results are saved in}

\item{save_dir}{is the directory to save the datasets.  Datasets can get quite large so choose a directory with plenty of free space.}

\item{GibbsSamps}{is TRUE if the inferential replicates are Gibbs samples and FALSE is the replicates are bootstrap samples}

\item{filteredgenenames}{is a character vector of all genenames that pass filtering}

\item{cntGene}{is the data.frame of counts and lengths for each sample saved by \code{\link{sumToGene}}}

\item{key}{is a data.frame with columns "Sample" (corresponding to te unique bioogical identifier for the analysis), "Condition" (giving the condition/treatment effect variables for the data),
and "Identifier", which should just be "Sample1", "Sample2", ... up to the number of rows of key.  This "Identifier" needs to be created like this even if
the observations don't correspond to unique biological samples.}

\item{curr_part_num}{is the current part number that is being run.  See the example in the sample code  (3)SaveNecessaryDatasetsForCompDTUReg.R}

\item{DRIMSeqFiltering}{is TRUE if DRIMSeq's filtering method is used.  Should be set to TRUE.}
}
\description{
Save the datasets that contain data across all inferential replicates and samples for a particular subset of genes
}
\details{
This function is used to save the necessary inferential replicate datasets. See the file (3)SaveNecessaryDatasetsForCompDTUReg.R for example code to work with this function.
}
