% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SaveGeneLevelFiles}
\alias{SaveGeneLevelFiles}
\title{Save a file for each gene containing all information necessary to run CompDTUReg}
\usage{
SaveGeneLevelFiles(
  dir1,
  GeneLevelFilesSaveDir,
  DRIMSeqFiltering = TRUE,
  useInferentialReplicates = TRUE,
  CorrectLowExpression = TRUE,
  CorrectLowExpressionParam = 0.01,
  GibbsSamps
)
}
\arguments{
\item{dir1}{is the directory where previous datasets are saved by \code{\link{sumToGene}} and \code{\link{DRIMSeqFilter}}}

\item{GeneLevelFilesSaveDir}{is the directory the gene level files will be saved to}

\item{DRIMSeqFiltering}{is an optional parameter indicating if DRIMSeqs filtering procedure is used.  Default is TRUE, FALSE is only provided as legacy option for
using the OtherGroups approach we considered and generally should not be used.}

\item{useInferentialReplicates}{is set to TRUE if inferential replicates are to be used in the analysis and FALSE otherwise}

\item{GibbsSamps}{is TRUE/FALSE specifying whether Gibbs samples of bootstrap samples are being used}
}
\description{
Save a file for each gene containing all information necessary to run CompDTUReg
}
