% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SaveInfRepDataAsRData}
\alias{SaveInfRepDataAsRData}
\title{Save the infRep data for all genes across each unique biological sample/replicate.}
\usage{
SaveInfRepDataAsRData(curr_samp, curr_file_loc, GibbsSamps = FALSE,
  countsFromAbundance = "no", direc_to_save_res = NULL)
}
\arguments{
\item{curr_samp}{Gives the current sample number in the form of "Sample1", "Sample2", etc.  Used in conjunction with key$Identifier, so needs to be "Sample1", "Sample2", etc
even if the observation does not correspond to a unique biological sample}

\item{curr_file_loc}{Gives the location of the .sf file for the Salmon quantification of the current Sample}

\item{GibbsSamps}{is TRUE/FALSE specifying whether Gibbs samples of bootstrap samples are being used}

\item{direc_to_save_res}{is an optional directory to save the sample specific dataset.  If unspecifying it will be saved to the current working directory.}
}
\description{
Save the infRep data for all genes across each unique biological sample/replicate.
}
