% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDTUReg.R
\name{CompDTUReg}
\alias{CompDTUReg}
\title{Implement the CompDTU and CompDTUme regression model}
\usage{
CompDTUReg(
  genename,
  Y = NULL,
  Group,
  runWithME = TRUE,
  YInfRep = NULL,
  mean.withinhat = NULL,
  extraPredictors = NULL
)
}
\arguments{
\item{genename}{The name of the current gene to run the method on.}

\item{Y}{corresponds is the ilr transformed matrix of response values for the non-inferential replicate data.  Matrix
should have a number of rows corresponding to the number of samples and a number of columns corresponding to the number of ilr coordinates
(which one less than the number transcripts remaining after fltering).  Set to NULL if you are running the model with measurement error, as it is not used.}

\item{Group}{A vector of condition assignments corresponding to the samples}

\item{runWithME}{is a T/F indicating whether the model should be run with measurement error or not (corresponding to CompDTU
and CompDTUme models respectively).  If runWithME is TRUE ensure YInfRep is non-NULL and if runWithME is FALSE ensure Y is non-NULL.}

\item{YInfRep}{corresponds is the ilr transformed matrix of response values for the inferential replicate data.  Matrix
should have a number of rows corresponding to the number of samples times the number of replicates and a number of columns corresponding to the number of ilr coordinates
(which one less than the number transcripts remaining after fltering).  Set to NULL if you are running the model without measurement error, as it is not used.}

\item{extraPredictors}{is an optional matrix of additional predictor values.  This should have one row per sample and one column per predictor.}
}
\value{
a data.frame containing the gene_id being used, p-value for the CompDTU/CompDTUme, and various information on
the current dataset being used
}
\description{
\code{CompDTUReg} runs the CompDTU and CompDTUme regression models for a specific gene.
}
