% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{DRIMSeqFilter}
\alias{DRIMSeqFilter}
\title{Filter data using filtering procedure built into \emph{DRIMSeq} via the \code{\link{dmFilter}} function.  Will automatically save
the filtered versions of the various datasets described in \code{\link{sumToGene}}}
\usage{
DRIMSeqFilter(failedinfRepsamps = NULL)
}
\arguments{
\item{failedinfRepsamps}{is an optional parameters that gives names of samples that had the infRep sampler fail.
This should not be needed, as newer versions of Salmon don't seem to have this issue.}
}
\description{
Filter data using filtering procedure built into \emph{DRIMSeq} via the \code{\link{dmFilter}} function.  Will automatically save
the filtered versions of the various datasets described in \code{\link{sumToGene}}
}
