% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepareData}
\alias{prepareData}
\title{Prepare raw adbundance/count/length data for later analysis}
\usage{
prepareData(
  abundance,
  counts,
  lengths,
  tx2gene,
  nsamp,
  key = NULL,
  infReps = "none",
  samps = NULL
)
}
\arguments{
\item{abundance}{is a dataframe with nsamp+1 columns, with names Sample1, Sample2, etc and a column for tx_id (that often comes from the rownames).  Rows are transcript level quantification estimates.  Column names should not include "TPM".}

\item{counts}{is a dataframe with nsamp+1 columns, with names Sample1, Sample2, etc and a column for tx_id (that often comes from the rownames).  Rows are transcript level quantification estimates. Column names should not include "Cnt".}

\item{tx2gene}{is a dataframe that matches transcripts to genes. Can be created by \code{\link{maketx2gene}}.}

\item{nsamp}{is the number of biological samples/replicates used in the analysis}

\item{key}{is a data.frame with columns "Sample" (corresponding to te unique bioogical identifier for the analysis), "Condition" (giving the condition/treatment effect variables for the data),
and "Identifier", which should just be "Sample1", "Sample2", ... up to the number of rows of key.  This "Identifier" needs to be created like this even if
the observations don't correspond to unique biological samples.}

\item{samps}{is an optional vector containing the sample names.  Need to specify this if sample names are not just paste0("Sample", 1:nsamp) without any missing.}

\item{length}{is a dataframe with nsamp+1 columns, with names Sample1, Sample2, etc and a column for tx_id (that often comes from the rownames).  Rows are transcript level effective length information. Column names should not include "Length".}

\item{CompMI}{is a TRUE/FALSE corresponding to whether datasets for the multiple imputation based analysis are being used.  This will add columns for transcripts that may be missing in the inferential replicates that were't missing in the non-inferential replicate data.  Default is FALSE.}
}
\value{
list of length 2 with the first element being the abundance data (abGeneTempF) and the second being the count data (cntGeneTempF) for use with \code{\link{generateData}}
}
\description{
\code{prepareData} generates abundance and count data to be used later, notably in \code{\link{generateData}}
}
