% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SaveWithinSubjCovMatrices}
\alias{SaveWithinSubjCovMatrices}
\title{Save the ilr-transformed within subject covariance matricies for the analysis}
\usage{
SaveWithinSubjCovMatrices(DRIMSeqFiltering = TRUE)
}
\arguments{
\item{DRIMSeqFiltering}{is an optional parameter indicating if DRIMSeqs filtering procedure is used.  Default is TRUE, FALSE is only provided as legacy option for
using the OtherGroups approach we considered and generally should not be used}
}
\description{
Save the ilr-transformed within subject covariance matricies for the analysis
}
