% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompDTUReg.R
\name{startCompDTUReg}
\alias{startCompDTUReg}
\title{\code{startCompDTUReg} start the compositional DTU regression model corresponding to a specific gene-level file.}
\usage{
startCompDTUReg(x, runWithME, extraPredictors = NULL)
}
\arguments{
\item{x}{The file path to a specific gene-level file}

\item{runWithME}{is a T/F indicating whether the model should be run with measurement error or not (corresponding to CompDTU
and CompDTUme models respectively).  If runWithME is TRUE ensure YInfRep is non-NULL and if runWithME is FALSE ensure Y is non-NULL.}

\item{extraPredictors}{is an optional matrix of additional predictor values.  This should have one row per sample and one column per perdictor.}
}
\description{
\code{startCompDTUReg} runs the compositional DTU regression models.  This is a helper function to run \code{\link{CompDTUReg}} from gene level files originating
from the file (4)RunCompositionalRegressions in the sample code for the package.
}
